+++ lib/src/roulette.dart
@@ -14,45 +14,46 @@
 
 import 'package:flutter/material.dart';
 
+import 'roulette_scope.dart';
 import 'roulette_style.dart';
 import 'roulette_controller.dart';
 import 'roulette_paint.dart';
 
 /// This is an animatable roulette widget.
 /// You need to present a [RouletteController] to controll this widget.
-class Roulette extends StatefulWidget {
+class Roulette extends StatelessWidget {
   const Roulette({
     Key? key,
-    required this.controller,
+    this.controller,
     this.style = const RouletteStyle(),
   }) : super(key: key);
 
   /// Controls the roulette.
-  final RouletteController controller;
+  final RouletteController? controller;
 
   /// The display style of the roulette.
   final RouletteStyle style;
 
   @override
-  State<Roulette> createState() => _RouletteState();
-}
-
-class _RouletteState extends State<Roulette> {
-  @override
-  void initState() {
-    widget.controller.addListener(() {
-      setState(() {});
-    });
-    super.initState();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return RoulettePaint(
-      key: widget.key,
-      animation: widget.controller.animation,
-      style: widget.style,
-      group: widget.controller.group,
+    final c = this.controller ?? RouletteScope.of(context);
+    assert(
+      c != null,
+      'no controller or scope controller of Roulette provided. '
+      'you should provide controller directly on Roulette widget or '
+      'specify it using RouletteScope',
+    );
+    final controller = c!;
+    return AnimatedBuilder(
+      animation: controller,
+      builder: (context, child) {
+        return RoulettePaint(
+          key: key,
+          animation: controller.animation,
+          style: style,
+          group: controller.group,
+        );
+      },
     );
   }
 }